% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{mapSenegal}
\alias{mapSenegal-package}
\alias{mapSenegal}
\title{Administrative Boundaries of Senegal}
\description{
The administrative boundaries of Senegal are provided at several
levels, including regions, departments, arrondissements and communes.
The Global Administrative Areas database, or `GADM`, is
the primary source for these layers. The dataset is complemented by the
incorporation of additional geographic layers, such as localities,
universities, roads, and health facility locations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mapSenegal/mapSenegal}
  \item Report bugs at \url{https://github.com/mapSenegal/mapSenegal/issues}
}

}
\author{
\strong{Maintainer}: Timothée Giraud \email{timothee.giraud@cnrs.fr} (\href{https://orcid.org/0000-0002-1932-3323}{ORCID})

Authors:
\itemize{
  \item Labaly Touré (\href{https://orcid.org/0009-0008-0939-2019}{ORCID})
}

Other contributors:
\itemize{
  \item Hugues Pecout (\href{https://orcid.org/0000-0002-0246-0954}{ORCID}) (Logo) [contributor]
  \item Robert J. Hijmans (\href{https://orcid.org/0000-0001-5872-2872}{ORCID}) (GADM database) [copyright holder]
}

}
