% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_arrondissements}
\alias{sn_arrondissements}
\title{Arrondissements}
\source{
Codes and geometries have been extracted from the GADM database v4.1
(\url{https://gadm.org/}). Geometries precision has been fixed to 1 meter.
}
\usage{
sn_arrondissements()
}
\value{
An sf object is returned.
}
\description{
Arrondissements of Senegal.
\itemize{
\item GID_1: identifier for regions
\item GID_2: identifier for departments
\item GID_3: unique identifier for arrondissements
\item NAME: name of the arrondissement
\item CC_3: unknown code
}
}
\examples{
ardt <- sn_arrondissements()
plot(ardt)
if(require("mapsf")){
  mf_map(ardt)
  mf_title("Arrondissements of Senegal")
}
}
