% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_communes}
\alias{sn_communes}
\title{Communes}
\source{
Codes and geometries have been extracted from the GADM database v4.1
(\url{https://gadm.org/}). Geometries precision has been fixed to 1 meter.
}
\usage{
sn_communes()
}
\value{
An sf object is returned.
}
\description{
Communes of Senegal.
\itemize{
\item GID_1: identifier for regions
\item GID_2: identifier for departments
\item GID_3: identifier for arrondissements
\item GID_4: identifier for communes
\item NAME: name of the arrondissement
\item CC_4: unknown code
}
}
\examples{
com <- sn_communes()
plot(com)
if (require("mapsf")) {
  mf_map(com)
  mf_title("Communes of Senegal")
}
}
