% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_country}
\alias{sn_country}
\title{Country}
\source{
Codes and geometries have been extracted from the GADM database v4.1
(\url{https://gadm.org/}). Geometries precision has been fixed to 1 meter.
}
\usage{
sn_country()
}
\value{
An sf object is returned.
}
\description{
Senegal, level-0 entity.
\itemize{
\item GID_0: unique identifier for countries
\item NAME: name of the country
}
}
\examples{
senegal <- sn_country()
plot(senegal)
if(require("mapsf")){
  mf_map(senegal)
  mf_title("Senegal")
}
}
