% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_departments}
\alias{sn_departments}
\title{Departments}
\source{
Codes and geometries have been extracted from the GADM database v4.1
(\url{https://gadm.org/}). Geometries precision has been fixed to 1 meter.
}
\usage{
sn_departments()
}
\value{
An sf object is returned.
}
\description{
Departments of Senegal, level-2 entities.
\itemize{
\item GID_1: identifier for regions
\item GID_2: unique identifier for departments
\item NAME: name of the department
\item CC_2: unknown code
\item HASC_2: hierarchical administrative subdivision codes for level-2 entities
\item POP_RGPH5: 2023 population
\item POP_2024: 2024 population
}
}
\examples{
dep <- sn_departments()
plot(dep)
if (require("mapsf")) {
  mf_map(dep)
  mf_title("Departments of Senegal")
}
}
