% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_health_facilities}
\alias{sn_health_facilities}
\title{Health facilities}
\source{
Direction des Travaux geographiques et de la cartographie, Senegal.
}
\usage{
sn_health_facilities()
}
\value{
An sf object is returned.
}
\description{
Health facilities of Senegal
\itemize{
\item NAME: health facility name
\item CC_1: unkown code
\item CC_2: unkown code
\item TYPE: health facility category
}
}
\examples{
hf <- sn_health_facilities()
plot(hf)
if (require("mapsf")) {
  mf_map(hf, var = "TYPE", type = "symb", cex = 1, pal = "Dark 3", add = FALSE)
  mf_title("Health Facilities of Senegal")
}
}
