% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_localities}
\alias{sn_localities}
\title{Localities}
\source{
Unknown source
}
\usage{
sn_localities()
}
\value{
An sf object is returned.
}
\description{
Localites of Senegal.
\itemize{
\item NAME: locality name
\item CC_1: unkown code
\item CC_2: unkown code
}
}
\examples{
loc <- sn_localities()
plot(loc)
if(require("mapsf")){
  mf_map(loc)
  mf_title("Localities of Senegal")
}
}
