% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_neighbors}
\alias{sn_neighbors}
\title{Neighboring countries}
\source{
Codes and geometries have been extracted from the GADM database v4.1
(\url{https://gadm.org/}). Geometries precision has been fixed to 1 meter.
Countries were cropped using a 100 km buffer around the Senegal bounding box.
}
\usage{
sn_neighbors()
}
\value{
An sf object is returned.
}
\description{
Neighboring countries of Senegal, level-0 entities.
\itemize{
\item GID_0: unique identifier for countries
\item NAME: name of the country
}
}
\examples{
neighbors <- sn_neighbors()
plot(neighbors)
if(require("mapsf")){
  mf_map(neighbors)
  mf_label(neighbors, "NAME", halo = TRUE)
  mf_title("Senegal and Neighboring Countries")
}
}
