% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_regions}
\alias{sn_regions}
\title{Regions}
\source{
Codes and geometries have been extracted from the GADM database v4.1
(\url{https://gadm.org/}). Geometries precision has been fixed to 1 meter.
}
\usage{
sn_regions()
}
\value{
An sf object is returned.
}
\description{
Regions of Senegal, level-1 entities.
\itemize{
\item GID_1: unique identifier for regions
\item NAME: name of the region
\item CC_1: unknown identifier
\item HASC_1: hierarchical administrative subdivision codes for level-1 entities
\item ISO: ISO codes for level-1 entities (ISO-3166-2)
\item POP_RGPH4: 20xx population
\item POP_RGPH5: 2023 population
}
}
\examples{
reg <- sn_regions()
plot(reg)
if(require("mapsf")){
  mf_map(reg)
  mf_title("Regions of Senegal")
}
}
