% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_roads}
\alias{sn_roads}
\title{Roads}
\source{
Unknown source
}
\usage{
sn_roads()
}
\value{
An sf object is returned.
}
\description{
National and departmental roads of Senegal.
\itemize{
\item NAME: road name
\item TYPE: road type ("N" for national roads, "D" for departmental roads)
}
}
\examples{
roads <- sn_roads()
plot(roads)
if(require("mapsf")){
  mf_map(roads, var = "TYPE", type = "typo", lwd = 2)
  mf_title("Roads of Senegal")
}
}
