% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_universities}
\alias{sn_universities}
\title{Universities}
\source{
Dataset gathered by Labaly Toure.
}
\usage{
sn_universities()
}
\value{
An sf object is returned.
}
\description{
Universities of Senegal.
\itemize{
\item NAME: university name
\item NB_FACULTIES: number of faculties
\item NB_STUDENTS: number of students
}
}
\examples{
univ <- sn_universities()
plot(univ)
if (require("mapsf")) {
  mf_map(univ, var = "NB_STUDENTS", type = "prop", add = FALSE)
  mf_title("Universities of Senegal")
}
}
