% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{mapSenegal}
\alias{mapSenegal-package}
\alias{mapSenegal}
\title{Administrative Boundaries of Senegal}
\description{
Administrative boundaries of Senegal at several levels (regions,
departements, arrondissements and communes), based on 'GDAM', are provided.
The package also gives access to localities, universities, roads and health
facilities locations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mapSenegal/mapSenegal}
  \item Report bugs at \url{https://github.com/mapSenegal/mapSenegal/issues}
}

}
\author{
\strong{Maintainer}: Timothée Giraud \email{timothee.giraud@cnrs.fr} (\href{https://orcid.org/0000-0002-1932-3323}{ORCID})

Authors:
\itemize{
  \item Labaly Touré (\href{https://orcid.org/0009-0008-0939-2019}{ORCID})
}

Other contributors:
\itemize{
  \item Hugues Pecout (\href{https://orcid.org/0000-0002-0246-0954}{ORCID}) (Logo) [contributor]
  \item Robert J. Hijmans (\href{https://orcid.org/0000-0001-5872-2872}{ORCID}) (GADM database) [copyright holder]
}

}
