% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sn_roads}
\alias{sn_roads}
\title{Roads}
\source{
Direction des Travaux geographiques et de la cartographie, Senegal.
}
\usage{
sn_roads()
}
\value{
An sf object is returned.
}
\description{
Roads of Senegal.
\itemize{
\item TYPE: road type. 1 for motorways, 2 for national roads, 3 for regional
roads, 4 for departmental roads, 5 for other roads.
\item NAME: road name
}
}
\examples{
roads <- sn_roads()
plot(roads)
if (require("mapsf")) {
  mf_map(roads, var = "TYPE", type = "typo", lwd = 2)
  mf_title("Roads of Senegal")
}
}
